+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v2.0.0 (in development)
 - Support Python 3.14
 - Drop support for Python 3.8 and 3.9
 - Changed `ParsedWheelFilename` from a namedtuple to a dataclass
+- **Breaking**: Renamed `ParsedWheelFilename` to `WheelFilename`
 
 v1.4.2 (2024-12-01)
 -------------------